rs_viewer updates
- rs_viewer control on/off fixes;
- Error checks;
- Laser 2 trackbars work together like ae/exposure trackbars;
- Frame and control windows into one.

+++ utilities/streamApp/camera_sub_system/include/CSSTypes.h
@@ -100,9 +100,9 @@ typedef unsigned FourCC;
 struct RsBuffer {
     void *buffer;
     uint32_t length;
-    uint8_t index;
+    uint32_t index;
 
-    RsBuffer (void* b, uint32_t s, uint8_t i)
+    RsBuffer (void* b, uint32_t s, uint32_t i)
     : buffer(b), length(s), index(i) {}
 };
